그래프 버튼 상태 zustand로 공유
