+++ client/web/package.json
@@ -45,7 +45,7 @@
     "shortid": "^2.2.15",
     "slick-carousel": "^1.8.1",
     "typescript": "~3.7.2",
-    "zustand": "^3.5.2"
+    "zustand": "^3.5.7"
   },
   "scripts": {
     "start": "cross-env PORT=3001 react-scripts --max_old_space_size=4096 start",
